+++ lib/models/drive/drive_models.dart
@@ -1,0 +1,5 @@
+/// Library for Microsoft Graph API Drive models
+library drive_models;
+
+export 'drive_model.dart';
+export 'drive_item_model.dart';